requirements.txt
Flask
torch
numpy
Pillow
scikit-learn
joblib
rasterio
opencv-python
gunicorn
